Fix: tratamento de erro pra caso o usuário adicione um produto sem especificar os atributos

+++ controller/controlador_produto.py
@@ -3,25 +3,26 @@ import json
 class ControladorProduto:
     @staticmethod
     def adiciona_produto(objeto) -> None:
-        # Abre o arquivo no modo de leitura e escrita
-        json_file = open('produtos.json', 'r+')
-        # Tenta ler o conteúdo do arquivo JSON existente
-        try:
-            dados = json.load(json_file)
-        except json.decoder.JSONDecodeError:
-        # Caso o arquivo esteja vazio ou inválido, inicializa com um dicionário vazio
-            dados = {}
-        id = objeto.get_id()
-        produto = {"tipo": objeto.get_tipo(), "cor": objeto.get_cor(), "tamanho": objeto.get_tamanho(), "preco": objeto.get_preco(), "quantidade": objeto.get_quantidade()}
-        dados[id] = produto  # Adiciona o novo produto ao dicionário
+        if objeto is not None:
+            # Abre o arquivo no modo de leitura e escrita
+            json_file = open('produtos.json', 'r+')
+            # Tenta ler o conteúdo do arquivo JSON existente
+            try:
+                dados = json.load(json_file)
+            except json.decoder.JSONDecodeError:
+            # Caso o arquivo esteja vazio ou inválido, inicializa com um dicionário vazio
+                dados = {}
+            id = objeto.get_id()
+            produto = {"tipo": objeto.get_tipo(), "cor": objeto.get_cor(), "tamanho": objeto.get_tamanho(), "preco": objeto.get_preco(), "quantidade": objeto.get_quantidade()}
+            dados[id] = produto  # Adiciona o novo produto ao dicionário
 
-        # Retorna ao início do arquivo para sobrescrever o conteúdo antigo com o novo
-        json_file.seek(0)
-        # Escreve o dicionário completo no arquivo JSON
-        json.dump(dados, json_file, indent=4)
-        # Trunca o restante do arquivo, caso o novo conteúdo seja menor que o antigo
-        json_file.truncate() 
-        json_file.close() # Fecha o arquivo
+            # Retorna ao início do arquivo para sobrescrever o conteúdo antigo com o novo
+            json_file.seek(0)
+            # Escreve o dicionário completo no arquivo JSON
+            json.dump(dados, json_file, indent=4)
+            # Trunca o restante do arquivo, caso o novo conteúdo seja menor que o antigo
+            json_file.truncate() 
+            json_file.close() # Fecha o arquivo
 
 
     @staticmethod
